requirements.txt
click<=8.2.0
cooler
duckdb
h5py
ibis-framework[duckdb]
loguru
more-itertools
natsort
numpy
pandas<=2.0.1
plotly>5.0.0,<=5.10.0
pyarrow>8.0.0,<=9.0.0
pybedtools
pyranges
pysam>0.15.0,<=0.19.1
pyyaml>=6.0
quarto
ray
seaborn
sh
snakemake
tqdm
trackhub
tracknado
ujson
xopen
xxhash
